feat(api): 添加对DONE常量的处理及temperature字段的支持
- 添加了DONE常量用于识别数据流结束，并在DeepSeekApi中添加打印语句以通知用户流结束。
- 在ChatCompletionRequest中添加temperature字段，用于控制生成文本的随机性。
- 重构了ChatCompletionRequest的创建逻辑，将消息分类为系统消息和用户消息，并返回Result类型以处理错误。
- 更新了Usage和EmptyUsage的实现，将返回的prompt_tokens和completion_tokens类型从u64改为u32。
- 添加了DefaultUsage结构体，实现了Usage trait，并在DeepSeekChatModel中默认使用此结构体。
- 重构了ChatModelObservationContext的request_options方法以返回引用。
- 修改了ModelRequest trait的options方法返回类型为Option<&dyn ModelOptions>。
- 在ObservationImpl的相关方法中，将into_box替换为into_boxed以保持一致性。
- 更新了tests模块中的测试用例以包含DONE常量的字节表示。

Closes #123

+++ next-web-framework/next-web-ai/src/chat/meta_data/empty_usage.rs
@@ -4,11 +4,11 @@ use crate::chat::meta_data::usage::Usage;
 pub struct EmptyUsage;
 
 impl Usage for EmptyUsage {
-    fn get_prompt_tokens(&self) -> u64 {
+    fn get_prompt_tokens(&self) -> u32 {
         0
     }
 
-    fn get_completion_tokens(&self) -> u64 {
+    fn get_completion_tokens(&self) -> u32 {
         0
     }
 }
